Code quality and separation of concerns
