+++ src/client/components/FileUpload/FileUpload.scss
@@ -1,3 +1,3 @@
-.progress-bar-container {
+.file-upload__progress-bar-container {
   height: 25px;
 }
