+++ src/client/pages/CountryHome/Repository/hooks/useClosePanel.ts
@@ -1,7 +1,6 @@
 import { useCallback } from 'react'
 
 import { useAppDispatch } from 'client/store'
-import { FileUploadActions } from 'client/store/ui/fileUpload'
 import { RepositoryActions } from 'client/store/ui/repository'
 
 type Returned = () => void
@@ -9,7 +8,6 @@ type Returned = () => void
 export const useClosePanel = (): Returned => {
   const dispatch = useAppDispatch()
   return useCallback<Returned>(() => {
-    dispatch(FileUploadActions.reset())
     dispatch(RepositoryActions.setRepositoryItem(undefined))
   }, [dispatch])
 }
